frontdoor: renaming and splitting the files
