+++ azurerm/internal/services/frontdoor/validate/frontdoor_name.go
@@ -12,11 +12,3 @@ func FrontDoorName(i interface{}, k string) (_ []string, errors []error) {
 	}
 	return nil, nil
 }
-
-func FrontDoorWAFName(i interface{}, k string) (_ []string, errors []error) {
-	if m, regexErrs := validate.RegExHelper(i, k, `(^[a-zA-Z])([\da-zA-Z]{0,127})$`); !m {
-		return nil, append(regexErrs, fmt.Errorf(`%q must be between 1 and 128 characters in length, must begin with a letter and may only contain letters and numbers.`, k))
-	}
-
-	return nil, nil
-}
